adding search by title support

+++ bot.py
@@ -87,43 +87,48 @@ async def on_message(message):
                     await message.channel.send(embed=embedVar)
 
                     print('The link was not valid')
-            else:
-                #split the message after the fisrt 'empty space'
-                msg = msg.split(" ",1)
 
-                # create a youtube serach link with our string
-                html = urllib.request.urlopen(f"https://www.youtube.com/results?search_query={msg}")
-                video_ids = re.findall(r"watch\?v=(\S{11})", html.read().decode())
+            else: 
 
-                # construct a new url from the videos id's we got back
-                url = "https://www.youtube.com/watch?v=" + video_ids[0]
+                embedVar = discord.Embed(title="Something went wrong :confused: \n\nIt looks like you sent more than one url's, please send one url at time.\n\nCheck out -help and -info commands.", color=0x0066ff)
+                await message.channel.send(embed=embedVar)
+        elif not url:
+            #split the message after the fisrt 'empty space'
+            msg = msg.split(" ",1)
 
-                mp3.song(url)
-                os.listdir()
+            # create a youtube serach link with our string
+            html = urllib.request.urlopen(f"https://www.youtube.com/results?search_query={msg}")
+            video_ids = re.findall(r"watch\?v=(\S{11})", html.read().decode())
 
-                # get all of the .mp3 file in this directory
-                for files in glob.glob('*.mp3'):
+            # construct a new url from the videos id's we got back
+            url = "https://www.youtube.com/watch?v=" + video_ids[0]
 
-                    # for each .mp3 file get the file size
-                    file_size = getsize(files)
-                    # convert the file size into an integer
-                    file_size = int(file_size)
-            
-                    # check if the file size is over 8000000 bytes (discord limit for non bosted server's)
-                    if file_size > 8000000:
-                        print('The file size is over 8MB...\n')
+            mp3.song(url)
+            os.listdir()
 
-                        embedVar = discord.Embed(title="Something went wrong :confused:\n\nTry sending a song that is under 7 minutes long, \nbecause of Discord's file size limit.\\Check out -help and -info commands.", color=0x0066ff)
-                        await message.channel.send(embed=embedVar)
+            # get all of the .mp3 file in this directory
+            for files in glob.glob('*.mp3'):
 
-                        os.remove(files)
-                        print('File was deleted...')
-                    else:                
-                        await message.channel.send(file=discord.File(files))
-                        print('File was sent...\n')
+                # for each .mp3 file get the file size
+                file_size = getsize(files)
+                # convert the file size into an integer
+                file_size = int(file_size)
+        
+                # check if the file size is over 8000000 bytes (discord limit for non bosted server's)
+                if file_size > 8000000:
+                    print('The file size is over 8MB...\n')
 
-                        os.remove(files)
-                        print('File was deleted...\n')
+                    embedVar = discord.Embed(title="Something went wrong :confused:\n\nTry sending a song that is under 7 minutes long, \nbecause of Discord's file size limit.\\Check out -help and -info commands.", color=0x0066ff)
+                    await message.channel.send(embed=embedVar)
+
+                    os.remove(files)
+                    print('File was deleted...')
+                else:                
+                    await message.channel.send(file=discord.File(files))
+                    print('File was sent...\n')
+
+                    os.remove(files)
+                    print('File was deleted...\n')
         else:
             embedVar = discord.Embed(title="Something wen't wrong check out -help to example's.", color=0x0066ff)
             await message.channel.send(embed=embedVar)
